--- src/shared/ui/Dropdown/DropdownLink/index.ts
@@ -1,1 +1,0 @@
-export { DropdownLink } from './DropdownLink';
