+++ src/shared/ui/Layout/Layout.tsx
@@ -1,29 +1,15 @@
 import { Link, Outlet } from 'react-router-dom';
 
-import { useLogoutMutation, useMeQuery } from '@/entities/auth/api/auth';
-import { BaseErrorResponse } from '@/entities/auth/api/types';
+import { useMeQuery } from '@/entities/auth/api/auth';
 import { IncubatorLogo } from '@/shared/assets/icons/IncubatorLogo';
-import { LogOutIcon } from '@/shared/assets/icons/LogOutIcon';
 import { Routes } from '@/shared/constants/routes';
-import { errorNotification } from '@/shared/lib/notifications';
 import { Button } from '@/shared/ui/Button';
+import { UserMenu } from '@/shared/ui/Layout/UserMenu/UserMenu';
 
 import s from './Layout.module.scss';
 
 export const Layout = () => {
-  const { isSuccess } = useMeQuery();
-
-  const [logout] = useLogoutMutation();
-
-  const logoutHandler = async () => {
-    try {
-      await logout();
-    } catch (e) {
-      const error = e as BaseErrorResponse;
-
-      errorNotification(error.data.message || 'Some error occurred');
-    }
-  };
+  const { data, isSuccess } = useMeQuery();
 
   return (
     <>
@@ -32,11 +18,7 @@ export const Layout = () => {
           <IncubatorLogo />
         </a>
         {isSuccess ? (
-          // <UserMenu data={data} />
-          <Button onClick={logoutHandler} variant="secondary">
-            <LogOutIcon />
-            Logout
-          </Button>
+          <UserMenu data={data} />
         ) : (
           <Button as={Link} to={Routes.LOGIN} variant="secondary">
             Sign in
