+++ src/shared/ui/Layout/UserMenu/UserMenu.module.scss
@@ -1,4 +1,4 @@
-.dropdownLink {
+.item {
   cursor: pointer;
 
   display: flex;
@@ -6,6 +6,7 @@
   align-items: center;
 
   width: 100%;
+  padding: 0;
 
   text-decoration: none;
